llm_backend/requirements.txt
# app/requirements.txt
fastapi==0.112.1
uvicorn==0.30.6
transformers==4.44.1
torch==2.4.0
pyyaml==6.0.1
langchain==0.3.0
langchain_core==0.3.0  
langchain-community==0.3.0
langchain-chroma==0.1.4
langgraph==0.2.26
chromadb==0.4.23
sentence_transformers==2.3.1
typing-extensions==4.10.0
typing-inspect==0.8.0
torchvision==0.19.0
torchaudio==2.4.0
pydantic>=1.7,<3.0